requirements.txt
click
pyyaml
fleet
jinja2
